+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+* `AsyncData<T>` now supports nullable T.
+
+### Breaking Changes
+
+* AsyncData.loading has removed the parameter value.
+
 ## 1.0.1
 
 * Optimize the code.